# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OrgAInise API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ai
    description: AI-powered operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: data
    description: CRUD operations for project data (authenticated users only).
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /ai/analyze-session:
    post:
      operationId: analyzeSession
      tags: [ai]
      summary: Analyze session notes
      description: Sends session notes and existing memory to OpenAI, returns suggested updates grouped by category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionAnalysisInput"
      responses:
        "200":
          description: AI-generated suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/generate-context:
    post:
      operationId: generateContextBlock
      tags: [ai]
      summary: Generate a context block
      description: Compresses saved memory into a clean context block for pasting into AI chats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextBlockInput"
      responses:
        "200":
          description: Generated context block
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextBlockResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'

  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.

  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.

  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.

  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  # ── Data CRUD ─────────────────────────────────────────────────────
  /projects:
    get:
      tags: [data]
      operationId: listProjects
      summary: List all projects for the authenticated user
      responses:
        '200':
          description: Projects listed.
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/DbProject'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [data]
      operationId: upsertProject
      summary: Create or update a project (upsert by id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DbProject'
      responses:
        '200':
          description: Project saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbProject'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /projects/{projectId}:
    delete:
      tags: [data]
      operationId: deleteProject
      summary: Delete a project and all its data
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not your project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /projects/{projectId}/memories:
    get:
      tags: [data]
      operationId: listMemories
      summary: List memories for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memories listed.
          content:
            application/json:
              schema:
                type: object
                required: [memories]
                properties:
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/DbMemoryItem'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not your project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /memories/{memoryId}:
    put:
      tags: [data]
      operationId: upsertMemory
      summary: Create or update a memory item (upsert by id)
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DbMemoryItem'
      responses:
        '200':
          description: Memory item saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbMemoryItem'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [data]
      operationId: deleteMemory
      summary: Delete a memory item
      parameters:
        - name: memoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not your memory item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Memory item not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /projects/{projectId}/history:
    get:
      tags: [data]
      operationId: listSessionHistory
      summary: List session history for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: History listed.
          content:
            application/json:
              schema:
                type: object
                required: [history]
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/DbSessionHistory'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [data]
      operationId: createSessionHistory
      summary: Save a session history entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DbSessionHistory'
      responses:
        '200':
          description: History entry saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DbSessionHistory'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /sync:
    post:
      tags: [data]
      operationId: syncData
      summary: Bulk upsert all user data from localStorage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBody'
      responses:
        '200':
          description: Sync complete.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResult'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    MemoryItemSummary:
      type: object
      required: [text, category, importanceLevel]
      properties:
        text:
          type: string
        category:
          type: string
        importanceLevel:
          type: string
          enum: [must-include, useful-context, archive-reference]

    SessionAnalysisInput:
      type: object
      required: [projectName, projectType, categories, existingMemory, sessionNotes]
      properties:
        projectName:
          type: string
        projectType:
          type: string
        categories:
          type: array
          items:
            type: string
        existingMemory:
          type: array
          items:
            $ref: "#/components/schemas/MemoryItemSummary"
        sessionNotes:
          type: string

    AiSuggestion:
      type: object
      required: [suggestedText, category, importanceLevel, reason]
      properties:
        suggestedText:
          type: string
        category:
          type: string
        importanceLevel:
          type: string
          enum: [must-include, useful-context, archive-reference]
        reason:
          type: string
        conflictNote:
          type: ["string", "null"]

    SessionAnalysisResult:
      type: object
      required: [suggestions]
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/AiSuggestion"

    ContextBlockInput:
      type: object
      required: [projectName, projectType, length, selectedCategories, memoryItems]
      properties:
        projectName:
          type: string
        projectType:
          type: string
        length:
          type: string
          enum: [short, medium, full]
        selectedCategories:
          type: array
          items:
            type: string
        includeArchive:
          type: boolean
        memoryItems:
          type: array
          items:
            $ref: "#/components/schemas/MemoryItemSummary"

    ContextBlockResult:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ── Data schemas ─────────────────────────────────────────────────
    DbProject:
      type: object
      required: [id, name, type, categories, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        categories:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    DbMemoryItem:
      type: object
      required: [id, projectId, text, category, importanceLevel, createdAt, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        text:
          type: string
        category:
          type: string
        importanceLevel:
          type: string
          enum: [must-include, useful-context, archive-reference]
        createdAt:
          type: string
        updatedAt:
          type: string

    DbSessionHistory:
      type: object
      required: [id, projectId, rawNotes, suggestions, approvedCount, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        rawNotes:
          type: string
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/AiSuggestion'
        approvedCount:
          type: integer
        createdAt:
          type: string

    SyncBody:
      type: object
      required: [projects, memories, history]
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/DbProject'
        memories:
          type: array
          items:
            $ref: '#/components/schemas/DbMemoryItem'
        history:
          type: array
          items:
            $ref: '#/components/schemas/DbSessionHistory'

    SyncResult:
      type: object
      required: [syncedProjects, syncedMemories, syncedHistory]
      properties:
        syncedProjects:
          type: integer
        syncedMemories:
          type: integer
        syncedHistory:
          type: integer