# lib/api-zod/src/generated/types/dbMemoryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrgAInise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DbMemoryItemImportanceLevel } from './dbMemoryItemImportanceLevel';

export interface DbMemoryItem {
  id: string;
  projectId: string;
  text: string;
  category: string;
  importanceLevel: DbMemoryItemImportanceLevel;
  createdAt: string;
  updatedAt: string;
}