# lib/api-zod/src/generated/types/dbProject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrgAInise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DbProject {
  id: string;
  name: string;
  type: string;
  categories: string[];
  createdAt: string;
  updatedAt: string;
}