# lib/api-zod/src/generated/types/dbSessionHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrgAInise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiSuggestion } from './aiSuggestion';

export interface DbSessionHistory {
  id: string;
  projectId: string;
  rawNotes: string;
  suggestions: AiSuggestion[];
  approvedCount: number;
  createdAt: string;
}