# lib/api-zod/src/generated/types/syncBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrgAInise API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DbMemoryItem } from './dbMemoryItem';
import type { DbProject } from './dbProject';
import type { DbSessionHistory } from './dbSessionHistory';

export interface SyncBody {
  projects: DbProject[];
  memories: DbMemoryItem[];
  history: DbSessionHistory[];
}