# lib/api-zod/src/generated/types/syncResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OrgAInise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SyncResult {
  syncedProjects: number;
  syncedMemories: number;
  syncedHistory: number;
}